+++ graphql-gradle-plugin-samples-Forum-client/src/main/java/com/graphql_java_generator/minimal_app/MinimalNonSpringApp.java
@@ -15,6 +15,7 @@ public class MinimalNonSpringApp {
 		QueryTypeExecutor queryExecutor = new QueryTypeExecutor("http://localhost:8180/graphql");
 
 		// A basic demo of input parameters
+		@SuppressWarnings("deprecation")
 		Date date = new Date(2019 - 1900, 12 - 1, 20);
 
 		// For this simple sample, we execute a direct query. But prepared queries are recommended.
@@ -24,5 +25,6 @@ public class MinimalNonSpringApp {
 						+ date + "'");
 		System.out
 				.println(queryExecutor.boards("{id name publiclyAvailable topics(since: &param){id}}", "param", date));
+		System.out.println("Normal end of the application");
 	}
 }